Extract StoragePage and Alert classes
